Further progress on roboticon UI redesign
